+++ apps/meteor/client/contexts/VideoConfContext.ts
@@ -3,7 +3,7 @@ import { createContext, useContext } from 'react';
 import type { Subscription } from 'use-subscription';
 import { useSubscription } from 'use-subscription';
 
-import type { DirectCallParams, ProviderCapabilities, CallPreferences, VideoConfManager } from '../lib/VideoConfManager';
+import type { DirectCallData, ProviderCapabilities, CallPreferences, VideoConfManager } from '../lib/VideoConfManager';
 
 export type VideoConfPopupPayload = {
 	id: string;
@@ -22,7 +22,7 @@ type VideoConfContextValue = {
 	rejectIncomingCall: (callId: string) => void;
 	abortCall: () => void;
 	setPreferences: (prefs: { mic?: boolean; cam?: boolean }) => void;
-	queryIncomingCalls: Subscription<DirectCallParams[]>;
+	queryIncomingCalls: Subscription<DirectCallData[]>;
 	queryRinging: Subscription<boolean>;
 	queryCalling: Subscription<boolean>;
 	queryCapabilities: Subscription<ProviderCapabilities>;
@@ -47,7 +47,7 @@ export const useVideoConfJoinCall = (): VideoConfContextValue['joinCall'] => use
 export const useVideoConfDismissCall = (): VideoConfContextValue['dismissCall'] => useVideoConfContext().dismissCall;
 export const useVideoConfAbortCall = (): VideoConfContextValue['abortCall'] => useVideoConfContext().abortCall;
 export const useVideoConfRejectIncomingCall = (): VideoConfContextValue['rejectIncomingCall'] => useVideoConfContext().rejectIncomingCall;
-export const useVideoConfIncomingCalls = (): DirectCallParams[] => {
+export const useVideoConfIncomingCalls = (): DirectCallData[] => {
 	const { queryIncomingCalls } = useVideoConfContext();
 	return useSubscription(queryIncomingCalls);
 };