+++ apps/meteor/client/views/omnichannel/directory/contacts/contextualBar/ContactEditWithData.js
@@ -1,4 +1,4 @@
-import { Box } from '@rocket.chat/fuselage';
+import { Box, ContextualbarContent } from '@rocket.chat/fuselage';
 import { useTranslation } from '@rocket.chat/ui-contexts';
 import React, { useMemo } from 'react';
 
@@ -16,7 +16,11 @@ function ContactEditWithData({ id, close }) {
 	} = useEndpointData('/v1/omnichannel/contact', { params: useMemo(() => ({ contactId: id }), [id]) });
 
 	if ([state].includes(AsyncStatePhase.LOADING)) {
-		return <FormSkeleton />;
+		return (
+			<ContextualbarContent>
+				<FormSkeleton />
+			</ContextualbarContent>
+		);
 	}
 
 	if (error || !data || !data.contact) {