+++ apps/meteor/server/startup/watchDb.ts
@@ -1,10 +1,10 @@
 import { api } from '@rocket.chat/core-services';
+import { Logger } from '@rocket.chat/logger';
 import { MongoInternals } from 'meteor/mongo';
 
 import { metrics } from '../../app/metrics/server/lib/metrics';
 import { DatabaseWatcher } from '../database/DatabaseWatcher';
 import { db } from '../database/utils';
-import { Logger } from '../lib/logger/Logger';
 import { SystemLogger } from '../lib/logger/system';
 import { initWatchers } from '../modules/watchers/watchers.module';