+++ packages/rest-typings/src/v1/omnichannel.ts
@@ -21,6 +21,10 @@ import type {
 	IOmnichannelServiceLevelAgreements,
 	ILivechatPriority,
 	LivechatDepartmentDTO,
+	ILivechatTriggerCondition,
+	ILivechatTriggerAction,
+	ReportResult,
+	ReportWithUnmatchingElements,
 } from '@rocket.chat/core-typings';
 import { ILivechatAgentStatus } from '@rocket.chat/core-typings';
 import Ajv from 'ajv';
@@ -2982,11 +2986,236 @@ const POSTomnichannelIntegrationsSchema = {
 
 export const isPOSTomnichannelIntegrations = ajv.compile<POSTomnichannelIntegrations>(POSTomnichannelIntegrationsSchema);
 
+type POSTLivechatTranscriptRequestParams = {
+	email: string;
+	subject: string;
+};
+
+const POSTLivechatTranscriptRequestParamsSchema = {
+	type: 'object',
+	properties: {
+		email: {
+			type: 'string',
+		},
+		subject: {
+			type: 'string',
+		},
+	},
+	required: ['email', 'subject'],
+	additionalProperties: false,
+};
+
+export const isPOSTLivechatTranscriptRequestParams = ajv.compile<POSTLivechatTranscriptRequestParams>(
+	POSTLivechatTranscriptRequestParamsSchema,
+);
+
+type POSTLivechatTriggersParams = {
+	name: string;
+	description: string;
+	enabled: boolean;
+	runOnce: boolean;
+	conditions: ILivechatTriggerCondition[];
+	actions: ILivechatTriggerAction[];
+	_id?: string;
+};
+
+const POSTLivechatTriggersParamsSchema = {
+	type: 'object',
+	properties: {
+		name: {
+			type: 'string',
+		},
+		description: {
+			type: 'string',
+		},
+		enabled: {
+			type: 'boolean',
+		},
+		runOnce: {
+			type: 'boolean',
+		},
+		conditions: {
+			type: 'array',
+			items: {
+				type: 'object',
+				properties: {
+					name: {
+						type: 'string',
+						enum: ['time-on-site', 'page-url', 'chat-opened-by-visitor', 'after-guest-registration'],
+					},
+					value: {
+						type: 'string',
+						nullable: true,
+					},
+				},
+				required: ['name', 'value'],
+				additionalProperties: false,
+			},
+			minItems: 1,
+		},
+		actions: {
+			type: 'array',
+			items: {
+				type: 'object',
+				properties: {
+					name: {
+						type: 'string',
+						enum: ['send-message'],
+					},
+					params: {
+						type: 'object',
+						nullable: true,
+						properties: {
+							sender: {
+								type: 'string',
+								enum: ['queue', 'custom'],
+							},
+							msg: {
+								type: 'string',
+							},
+							name: {
+								type: 'string',
+								nullable: true,
+							},
+						},
+						required: ['sender', 'msg'],
+						additionalProperties: false,
+					},
+				},
+				required: ['name'],
+				additionalProperties: false,
+			},
+			minItems: 1,
+		},
+		_id: {
+			type: 'string',
+			nullable: true,
+		},
+	},
+	required: ['name', 'description', 'enabled', 'runOnce', 'conditions', 'actions'],
+	additionalProperties: false,
+};
+
+export const isPOSTLivechatTriggersParams = ajv.compile<POSTLivechatTriggersParams>(POSTLivechatTriggersParamsSchema);
+
+type POSTLivechatAppearanceParams = {
+	_id: string;
+	value: string | boolean | number;
+}[];
+
+const POSTLivechatAppearanceParamsSchema = {
+	type: 'array',
+	items: {
+		type: 'object',
+		properties: {
+			_id: {
+				type: 'string',
+			},
+			value: {
+				type: ['string', 'boolean', 'number'],
+			},
+		},
+		required: ['_id', 'value'],
+		additionalProperties: false,
+	},
+	minItems: 1,
+};
+
+export const isPOSTLivechatAppearanceParams = ajv.compile<POSTLivechatAppearanceParams>(POSTLivechatAppearanceParamsSchema);
+
+type GETDashboardConversationsByType = {
+	start: string;
+	end: string;
+	sort?: string;
+};
+
+const GETDashboardConversationsByTypeSchema = {
+	type: 'object',
+	properties: {
+		start: {
+			type: 'string',
+		},
+		end: {
+			type: 'string',
+		},
+		sort: {
+			type: 'string',
+		},
+	},
+	required: ['start', 'end'],
+	additionalProperties: false,
+};
+
+export const isGETDashboardConversationsByType = ajv.compile<GETDashboardConversationsByType>(GETDashboardConversationsByTypeSchema);
+
+type LivechatAnalyticsAgentOverviewProps = {
+	name: string;
+	from: string;
+	to: string;
+	departmentId?: string;
+};
+
+const LivechatAnalyticsAgentOverviewPropsSchema = {
+	type: 'object',
+	properties: {
+		name: {
+			type: 'string',
+		},
+		from: {
+			type: 'string',
+		},
+		to: {
+			type: 'string',
+		},
+		departmentId: {
+			type: 'string',
+			nullable: true,
+		},
+	},
+	required: ['name', 'from', 'to'],
+	additionalProperties: false,
+};
+
+export const isLivechatAnalyticsAgentOverviewProps = ajv.compile<LivechatAnalyticsAgentOverviewProps>(
+	LivechatAnalyticsAgentOverviewPropsSchema,
+);
+
+type LivechatAnalyticsOverviewProps = {
+	name: string;
+	from: string;
+	to: string;
+	departmentId?: string;
+};
+
+const LivechatAnalyticsOverviewPropsSchema = {
+	type: 'object',
+	properties: {
+		name: {
+			type: 'string',
+		},
+		from: {
+			type: 'string',
+		},
+		to: {
+			type: 'string',
+		},
+		departmentId: {
+			type: 'string',
+			nullable: true,
+		},
+	},
+	required: ['name', 'from', 'to'],
+	additionalProperties: false,
+};
+
+export const isLivechatAnalyticsOverviewProps = ajv.compile<LivechatAnalyticsOverviewProps>(LivechatAnalyticsOverviewPropsSchema);
+
 export type OmnichannelEndpoints = {
 	'/v1/livechat/appearance': {
 		GET: () => {
 			appearance: ISetting[];
 		};
+		POST: (params: POSTLivechatAppearanceParams) => void;
 	};
 	'/v1/livechat/visitors.info': {
 		GET: (params: LivechatVisitorsInfo) => {
@@ -3114,7 +3343,7 @@ export type OmnichannelEndpoints = {
 
 	'/v1/livechat/users/:type': {
 		GET: (params: LivechatUsersManagerGETProps) => PaginatedResult<{
-			users: ILivechatAgent[];
+			users: (ILivechatAgent & { departments: string[] })[];
 		}>;
 		POST: (params: POSTLivechatUsersTypeProps) => { success: boolean };
 	};
@@ -3150,7 +3379,7 @@ export type OmnichannelEndpoints = {
 
 	'/v1/livechat/users/agent': {
 		GET: (params: PaginatedRequest<{ text?: string }>) => PaginatedResult<{
-			users: ILivechatAgent[];
+			users: (ILivechatAgent & { departments: string[] })[];
 		}>;
 		POST: (params: LivechatUsersManagerPOSTProps) => { success: boolean };
 	};
@@ -3287,6 +3516,10 @@ export type OmnichannelEndpoints = {
 	'/v1/livechat/transcript': {
 		POST: (params: POSTLivechatTranscriptParams) => { message: string };
 	};
+	'/v1/livechat/transcript/:rid': {
+		DELETE: () => void;
+		POST: (params: POSTLivechatTranscriptRequestParams) => void;
+	};
 	'/v1/livechat/offline.message': {
 		POST: (params: POSTLivechatOfflineMessageParams) => { message: string };
 	};
@@ -3351,6 +3584,7 @@ export type OmnichannelEndpoints = {
 	};
 	'/v1/livechat/triggers': {
 		GET: (params: GETLivechatTriggersParams) => PaginatedResult<{ triggers: WithId<ILivechatTrigger>[] }>;
+		POST: (params: POSTLivechatTriggersParams) => void;
 	};
 	'/v1/livechat/triggers/:_id': {
 		GET: () => { trigger: ILivechatTrigger | null };
@@ -3443,6 +3677,18 @@ export type OmnichannelEndpoints = {
 	};
 	'/v1/livechat/rooms/filters': {
 		GET: () => { filters: IOmnichannelRoom['source'][] };
+	};
+	'/v1/livechat/analytics/agent-overview': {
+		GET: (params: LivechatAnalyticsAgentOverviewProps) => {
+			head: { name: string }[];
+			data: { name: string; value: number }[];
+		};
+	};
+	'/v1/livechat/analytics/overview': {
+		GET: (params: LivechatAnalyticsOverviewProps) => {
+			title: string;
+			value: string | number;
+		}[];
 	};
 } & {
 	// EE
@@ -3542,5 +3788,20 @@ export type OmnichannelEndpoints = {
 	};
 	'/v1/livechat/inquiry.setSLA': {
 		PUT: (params: { roomId: string; sla: string }) => void;
+	};
+	'/v1/livechat/analytics/dashboards/conversations-by-source': {
+		GET: (params: GETDashboardConversationsByType) => ReportResult;
+	};
+	'/v1/livechat/analytics/dashboards/conversations-by-status': {
+		GET: (params: GETDashboardConversationsByType) => ReportResult;
+	};
+	'/v1/livechat/analytics/dashboards/conversations-by-department': {
+		GET: (params: GETDashboardConversationsByType) => ReportWithUnmatchingElements;
+	};
+	'/v1/livechat/analytics/dashboards/conversations-by-tags': {
+		GET: (params: GETDashboardConversationsByType) => ReportWithUnmatchingElements;
+	};
+	'/v1/livechat/analytics/dashboards/conversations-by-agent': {
+		GET: (params: GETDashboardConversationsByType) => ReportWithUnmatchingElements;
 	};
 };