+++ ee/apps/account-service/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @rocket.chat/account-service
 
+## 0.2.2
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.3.2
+- @rocket.chat/rest-typings@6.3.2
+- @rocket.chat/core-services@0.1.2
+- @rocket.chat/model-typings@0.0.8
+- @rocket.chat/models@0.0.8
+
+## 0.2.1
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.3.1
+- @rocket.chat/rest-typings@6.3.1
+- @rocket.chat/core-services@0.1.1
+- @rocket.chat/model-typings@0.0.7
+- @rocket.chat/models@0.0.7
+
 ## 0.2.0
 
 ### Minor Changes