+++ ee/packages/api-client/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @rocket.chat/api-client
 
+## 0.1.2
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.3.2
+- @rocket.chat/rest-typings@6.3.2
+
+## 0.1.1
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.3.1
+- @rocket.chat/rest-typings@6.3.1
+
 ## 0.1.0
 
 ### Minor Changes